Move LogCommandHandler back to common
