--- src/background_worker/go.mod
@@ -1,7 +1,0 @@
-module background_worker
-
-go 1.20
-
-replace btcRate/common => ../common
-
-require btcRate/common v0.0.0-00010101000000-000000000000
